app/backend/requirements.txt
fastapi[all]
diffusers
transformers
safetensors
omegaconf
pytorch_lightning